herbarium-sheet-plant-organ-detection/requirements.txt
kafka-python==2.0.2
requests==2.27.1
Pillow==9.1.0
cloudevents==1.2.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0
torchvision==0.12.0
--find-links https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.9/index.html
detectron2
